refactor how names are formatted.
A value name may be a label or a name,  A label contains '%v' in its name.
Labels format the value equal to or above the named value.

Seperator/delimiters
Added with a value equal to or greater than a number which requires a preceeding delimiter if there are numbers preceeding it.